Optional prevMinter for ClubMinter.
Add an optional new constructor argument to ClubMinter, which can be used
to specify a "previous" minter contract.  Shares minted in the previous
one will be counted towards the cap.

With this, it is possible to replace / supersede an existing minter
contract with a new one and still keep the tracking of minted shares
in sync between contract and GSP.