Add getClubsSlice to PackSale.
This adds two new view methods to PackSale, getNumClubs and getClubsSlice.
With them, it is possible to retrieve the list of clubs in the contract
in batches.  This helps with tier 5, where getAllClubs() cannot be executed
as the result is too large for one call (but splitting it into two calls
would be fine).