Do not mark "quote" functions as view.
This removes the "view" classification of the functions that quote a swap
in the SwapProvider and calling places in AutoConvert or SwappingPackSale.

Uniswap v3 has its quoting not marked as view, so we need that to support
using Uniswap v3 as well for the SwapProvider.  These functions will still not
modify state in a meaningful way, and will only be called off-chain anyway.